requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
protobuf==3.19.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
